Game manager refracture
Refractured the game manager so that it doesn't depend on UI elements. Moved the Text, slider, and other related UI code over to the In Game UI Manager to do this.
