+++ Assets/Scripts/UI/InGameUIManager.cs
@@ -1,8 +1,10 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Security.Cryptography.X509Certificates;
+using TMPro;
 using UnityEngine;
 using UnityEngine.InputSystem;
+using UnityEngine.UI;
 
 public class InGameUIManager : MonoBehaviour
 {
@@ -11,7 +13,10 @@ public class InGameUIManager : MonoBehaviour
     [SerializeField] GameObject _settingsUI;
     [SerializeField] GameObject _confirmationUI;
     [SerializeField] GameObject _gameOverUI;
-    [SerializeField] PlayerPrefSO masterSO, musicSO, soundSO;
+    [SerializeField] PlayerPrefSO masterSO, musicSO, soundSO, _BestTimePlayerPref;
+    [SerializeField] private TextMeshProUGUI endTimeText, bestTimeText, currTimeText;
+    [SerializeField] private Slider scoreSlider;
+    [SerializeField] private GameObject scoreColor;
 
     void Start()
     {
@@ -20,6 +25,14 @@ public class InGameUIManager : MonoBehaviour
         _settingsUI.SetActive(false);
         _confirmationUI.SetActive(false);
         _gameOverUI.SetActive(false);
+
+        if (currTimeText)
+        {
+            InvokeRepeating("Timer", 1f, _gameManager.TimePerSecond);
+        } else
+        {
+            Debug.LogWarning("Missing UI Reference: Timer");
+        }
     }
 
     private void OnEnable()
@@ -120,5 +133,50 @@ public class InGameUIManager : MonoBehaviour
         PlayerPrefs.DeleteKey(masterSO.currKey.ToString());
         PlayerPrefs.DeleteKey(musicSO.currKey.ToString());
         PlayerPrefs.DeleteKey(soundSO.currKey.ToString());
+    }
+
+    /// <summary>
+    /// Timer for counting how much time has passed. Also records best time if it is less than current time.
+    /// </summary>
+    public void Timer()
+    {
+        _gameManager.CurrentTime++;
+        currTimeText.text = "Time Past: " + _gameManager.CurrentTime;
+
+        if (_gameManager.Score >= _gameManager.TargetScore)
+        {
+            _gameManager.BestTime = _gameManager.CurrentTime;
+
+            PlayerPrefs.SetInt(_BestTimePlayerPref.currKey.ToString(), _gameManager.BestTime);
+            PlayerPrefs.Save();
+
+            CancelInvoke("Timer");
+
+            endTimeText.text = string.Format("Final Time: {0}", _gameManager.CurrentTime);
+            bestTimeText.text = string.Format("Best Time: {0}", _gameManager.BestTime);
+
+            _gameManager.OnGameEnd.Invoke();
+        }
+    }
+
+    public void ChangeProgressBar()
+    {
+        if (scoreSlider)
+        {
+            scoreSlider.value = _gameManager.Score / _gameManager.TargetScore;
+
+            if (scoreSlider.value < 0.25f)
+                scoreColor.GetComponent<Image>().color = Color.red;
+            else if (scoreSlider.value > 0.25f && scoreSlider.value < 0.75f)
+                scoreColor.GetComponent<Image>().color = Color.yellow;
+            else if (scoreSlider.value > 0.75f && scoreSlider.value < 0.99f)
+                scoreColor.GetComponent<Image>().color = Color.green;
+            else
+                scoreColor.GetComponent<Image>().color = Color.blue;
+        }
+        else
+        {
+            Debug.LogWarning("Missing UI Reference: Score Slider");
+        }
     }
 }
